Make sure all resource data are accessible by reading every byte
Data could not be accessible if the .NET image has been loaded by the OS
loader, and there are gaps between sections with no memory. An
AccessViolationException will be thrown in that case if anyone tries to
access the memory. This should only happen if the file has been
obfuscated and the obfuscator added invalid resources.