Generic[M]Var both now know which GenericParam they belong to
FullNameCreator has been updated to use the new GenericParam if it's
available so MethodDef/FieldDef/etc's FullName and other string properties
may now show the name of the GenericParam instead of !0 or !!3.
The following are not cached if they contain any generic parameters:
	InterfaceImpl
	MemberRef
	StandAloneSig
	TypeSpec
	MethodSpec
	GenericParamConstraint
This means that eg. calling ResolveTypeSpec(rid) twice with the same input
may return two different values.