Fix bugs discovered when the code was tested.
The test was performed by loading `Microsoft.VisualBasic.Compatibility.dll` extracted from the installation media of VS.NET 2002, reading the resources, writing them, and verifying they are ok.

The Regex for checking the resource reader class had to be updated as for some unknown to me reason the version was not included in the name.

The code for comparing the type names also had to be updated to seperate the type name from teh assembly name and version.